+++ mobile/app/src/main/java/com/example/colorimagemobile/ui/home/fragments/gallery/views/SelectionView.kt
@@ -105,8 +105,8 @@ class SelectionView(context: Context?): CanvasView(context) {
                 val id = DrawingObjectManager.getUuid(selectedShapeIndex) ?: return
                 this.translateData = TranslateData(
                     id = id,
-                    deltaX = 0,
-                    deltaY = 0
+                    deltaX = 0f,
+                    deltaY = 0f
                 )
                 DrawingSocketService.sendTransformSelectionCommand(translateData!!, "Translation")
                 break
@@ -139,12 +139,12 @@ class SelectionView(context: Context?): CanvasView(context) {
             val id = DrawingObjectManager.getUuid(selectedShapeIndex) ?: return
             this.translateData = TranslateData(
                 id = id,
-                deltaX = (translateData?.deltaX ?: 0) + dx.toInt(),
-                deltaY = (translateData?.deltaY ?: 0) + dy.toInt()
+                deltaX = (translateData?.deltaX ?: 0f) + dx,
+                deltaY = (translateData?.deltaY ?: 0f) + dy
             )
             DrawingSocketService.sendTransformSelectionCommand(translateData!!, "Translation")
             translationCommand = TranslateCommand(translateData!!)
-            translationCommand!!.setTransformation(dx.toInt(), dy.toInt())
+            translationCommand!!.setTransformation(dx, dy)
             translationCommand!!.execute()
 
             SelectionService.resetBoundingBox()
